Add files via upload
special cases for unary
implement the for loop of the splitting into the usual formatting for making a for loop - so i can make variable - so i can check 'current' and 'next' token. instead of 'last' and 'current' token.    so i can put the - in the correct spot instead of where the second -//+ was